requirements.txt
asgiref==3.8.1
Django==5.1.2
django-environ==0.11.2
django-unfold==0.43.0

drf-yasg==1.21.8
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

inflection==0.5.1
packaging==24.1
pillow==11.0.0
pre-commit==3.7.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
ruff==0.7.2
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
django-lifecycle==1.2.4
django-filter==24.3
django-jalali-date==1.1.3
celery==5.4.0
redis==5.2.0
kavenegar==1.1.2
ipython==8.29.0
elastic-apm==6.23.0
django-leaflet==0.31.0
django_resized==1.0.3
psycopg2
django-mptt==0.16.0
